requirements.txt
psutil>=5.8.0
Flask>=2.0.0
pyyaml>=6.0
watchdog>=2.1.0
gunicorn>=20.1.0
requests>=2.25.0
